requirements.txt
Flask
requests
pytest
bleak
pycryptodome
Pillow